+++ src/dtos/media.dto.ts
@@ -1,0 +1,4 @@
+import { OmitType } from '@nestjs/mapped-types';
+import { Media } from '@app/db-mongoose/media/media.entity';
+
+export class CreateUserDto extends OmitType(Media, ['_id', 'createdAt', 'updatedAt'] as const) {}