+++ src/dtos/user.dto.ts
@@ -1,0 +1,4 @@
+import { OmitType } from '@nestjs/mapped-types';
+import { User } from '@app/db-mongoose/user/user.entity';
+
+export class CreateUserDto extends OmitType(User, ['createdAt', 'updatedAt'] as const) {}